refactor(rig-agent): engine.rs / runner.rs / telemetry.rs
