+++ crates/rig-agent/src/agent/runner.rs
@@ -24,47 +24,24 @@
 //! # }
 //! ```
 
-use std::sync::{
-    Arc, Mutex,
-    atomic::{AtomicU64, Ordering},
-};
+use std::sync::{Arc, Mutex};
 
 use futures::StreamExt;
-use tracing::{Instrument, info_span, span::Id};
 
 use super::{
     ModelHandle,
-    completion::{Agent, AgentConfig, PreparedCompletionRequest},
-    hook::{
-        AgentHook, CompletionCall, CompletionCallAction,
-        CompletionResponse as CompletionResponseEvent, HookContext, HookStack,
-        InvalidToolCallAction, ModelTurnAction, ModelTurnFinished, ObservationAction, RequestPatch,
-        ToolCall as ToolCallEvent, ToolCallAction, ToolResultAction, ToolResultEvent,
-    },
-    prompt_request::{
-        PromptResponse,
-        streaming::{
-            DriveItem, DriveStream, MultiTurnStreamItem, StreamingError, TurnSource, drive_agent,
-            drive_tool_calls, streaming_error_into_prompt,
-        },
-        tool_result_output,
-    },
-    run::{AgentRun, ModelTurn, ModelTurnOutcome, PendingToolCall},
+    completion::{Agent, AgentConfig},
+    engine::{DriveItem, UnaryTurnSource, drive_agent, streaming_error_into_prompt},
+    hook::AgentHook,
+    prompt_request::PromptResponse,
+    run::AgentRun,
+    telemetry::acquire_agent_span,
 };
-use rig_core::{
-    completion::{FinishReason, ResponseIdentity},
-    memory::ConversationMemory,
-    message::{AssistantContent, ToolCall, ToolChoice, UserContent},
-    telemetry::SpanCombinator,
-};
+use rig_core::{memory::ConversationMemory, message::ToolChoice};
 
 use crate::{
     completion::{CompletionError, CompletionModel, Document, Message, PromptError, Usage},
-    json_utils,
-    tool::{
-        ToolContext, ToolDispatch, ToolResult,
-        server::{ToolRegistrySnapshot, ToolServerHandle},
-    },
+    tool::{ToolContext, server::ToolServerHandle},
 };
 
 use super::UNKNOWN_AGENT_NAME;
@@ -74,123 +51,6 @@ pub(crate) enum UnhandledInvalidToolCallPolicy {
     #[default]
     Fail,
     IgnoreForExtractor,
-}
-
-/// Build the per-turn `chat` span shared by both turn sources.
-///
-/// The span *name* must be a string literal — `tracing` bakes it into static
-/// metadata — so this is a macro parameterized by the name rather than a
-/// function (the two surfaces keep distinct names, `chat` vs `chat_streaming`,
-/// which dashboards split on). The matching operation value is passed with the
-/// name; every other field is identical across the two surfaces, so it lives
-/// here once instead of being copy-pasted into each `TurnSource::open_chat_span`.
-macro_rules! build_chat_span {
-    ($runner:expr, $effective_preamble:expr, $name:literal, $operation:literal) => {{
-        let system_instructions = $crate::core::telemetry::system_instructions_json(
-            $effective_preamble,
-            $runner.config.record_telemetry_content,
-        );
-        // The core macro is the single source of the completion-parent
-        // contract (marker + required fields); only the agent-specific field
-        // is declared here.
-        $crate::core::telemetry::completion_parent_span!(
-            target: "rig::agent_chat",
-            name: $name,
-            operation: $operation,
-            system_instructions: system_instructions.as_deref(),
-            gen_ai.agent.name = $runner.agent_name_or_default(),
-        )
-    }};
-}
-pub(crate) use build_chat_span;
-
-/// Convert an observe-only action into an optional stop reason.
-pub(crate) fn observe_action(action: ObservationAction) -> Option<String> {
-    match action {
-        ObservationAction::Continue => None,
-        ObservationAction::Stop(reason) => Some(reason),
-    }
-}
-
-/// Resolved outcome of the shared, medium-neutral model-turn hook.
-pub(crate) enum ModelTurnDecision {
-    /// Accept the turn and advance normally.
-    Advance,
-    /// The turn was rejected and the run is ready to issue another model call.
-    Retried,
-    /// Stop the run with the supplied reason.
-    Terminate(String),
-}
-
-/// One attempt's assembled response, as both drivers hand it to
-/// [`settle_model_turn`]: the unary response under `run()`, the finished
-/// stream under `stream()`. Every field is this attempt's own — a retry
-/// re-enters with a fresh response, so a stale attempt's ids or payload can
-/// never be attributed here.
-pub(crate) struct AssembledTurn<'a> {
-    pub(crate) prompt: &'a Message,
-    pub(crate) content: &'a Vec<AssistantContent>,
-    pub(crate) usage: Usage,
-    pub(crate) identity: &'a ResponseIdentity,
-    pub(crate) finish_reason: Option<&'a FinishReason>,
-    /// The cap this attempt was prepared with, completion-call patches
-    /// included; read off the prepared request, never the agent config.
-    pub(crate) max_tokens: Option<u64>,
-    pub(crate) raw: &'a serde_json::Value,
-}
-
-/// Settle a parked model turn: fire [`AgentHook::on_completion_response`]
-/// (observe-only; a stop terminates), then
-/// [`AgentHook::on_model_turn_finished`] and apply its action to the sans-IO
-/// run. Both drivers call this once per accepted attempt, so retry history,
-/// tool-turn rejection, and state transitions cannot diverge by medium. The
-/// callers own what happens next: the blocking driver records the accepted
-/// turn's telemetry; the streaming driver additionally surfaces or discards
-/// the buffered provisional `Final`.
-pub(crate) async fn settle_model_turn(
-    hooks: &HookStack,
-    hook_ctx: &HookContext,
-    run: &mut AgentRun,
-    turn: AssembledTurn<'_>,
-) -> Result<ModelTurnDecision, PromptError> {
-    if let Some(reason) = observe_action(
-        hooks
-            .on_completion_response(
-                hook_ctx,
-                CompletionResponseEvent {
-                    prompt: turn.prompt,
-                    content: turn.content,
-                    usage: turn.usage,
-                    identity: turn.identity,
-                    raw: turn.raw,
-                },
-            )
-            .await,
-    ) {
-        return Ok(ModelTurnDecision::Terminate(reason));
-    }
-    let action = hooks
-        .on_model_turn_finished(
-            hook_ctx,
-            ModelTurnFinished {
-                turn: hook_ctx.turn(),
-                content: turn.content,
-                usage: turn.usage,
-                identity: turn.identity,
-                finish_reason: turn.finish_reason,
-                max_tokens: turn.max_tokens,
-                raw: turn.raw,
-            },
-        )
-        .await;
-    match action {
-        ModelTurnAction::Continue => Ok(ModelTurnDecision::Advance),
-        ModelTurnAction::Retry(request) => {
-            run.retry_model_turn(request)?;
-            Ok(ModelTurnDecision::Retried)
-        }
-        ModelTurnAction::Stop(reason) => Ok(ModelTurnDecision::Terminate(reason)),
-    }
 }
 
 /// A hook-aware driver over [`AgentRun`].
@@ -534,539 +394,6 @@ pub(crate) fn build_agent_run(
     AgentRun::from_spec(&spec, prompt, history)
 }
 
-/// Build (or adopt) the top-level `invoke_agent` span for a run, shared by the
-/// blocking and streaming drivers so the run-level span shape is defined once.
-///
-/// Returns the span plus whether it was newly created. When the caller is
-/// already inside a span we adopt it and report `false`, so the driver can avoid
-/// recording run-level usage onto a span it does not own (see the
-/// `created_agent_span` guard in both drivers' `Done` handling).
-pub(crate) fn acquire_agent_span(
-    agent_name: &str,
-    preamble: Option<&str>,
-    record_content: bool,
-) -> (tracing::Span, bool) {
-    if tracing::Span::current().is_disabled() {
-        let system_instructions =
-            rig_core::telemetry::system_instructions_json(preamble, record_content);
-        let span = info_span!(
-            "invoke_agent",
-            gen_ai.operation.name = "invoke_agent",
-            gen_ai.agent.name = agent_name,
-            gen_ai.system_instructions = system_instructions.as_deref(),
-            gen_ai.prompt = tracing::field::Empty,
-            gen_ai.completion = tracing::field::Empty,
-            gen_ai.usage.input_tokens = tracing::field::Empty,
-            gen_ai.usage.output_tokens = tracing::field::Empty,
-            gen_ai.usage.cache_read.input_tokens = tracing::field::Empty,
-            gen_ai.usage.cache_creation.input_tokens = tracing::field::Empty,
-            gen_ai.usage.tool_use_prompt_tokens = tracing::field::Empty,
-            gen_ai.usage.reasoning_tokens = tracing::field::Empty,
-        );
-        (span, true)
-    } else {
-        (tracing::Span::current(), false)
-    }
-}
-
-/// Outcome of firing the `CompletionCall` hook for a turn.
-pub(crate) enum CompletionCallOutcome {
-    /// Proceed, optionally applying a per-turn request patch (the merged patch
-    /// from every hook that contributed one).
-    Proceed(Option<RequestPatch>),
-    /// Terminate the run with this reason.
-    Terminate(String),
-}
-
-/// Fire the event-specific completion-call hook for a turn.
-pub(crate) async fn resolve_completion_call(
-    hooks: &HookStack,
-    ctx: &HookContext,
-    prompt: &Message,
-    history: &[Message],
-    turn: usize,
-) -> CompletionCallOutcome {
-    match hooks
-        .on_completion_call(
-            ctx,
-            CompletionCall {
-                prompt,
-                history,
-                turn,
-            },
-        )
-        .await
-    {
-        CompletionCallAction::Stop(reason) => CompletionCallOutcome::Terminate(reason),
-        CompletionCallAction::Patch(patch) => CompletionCallOutcome::Proceed(Some(patch)),
-        CompletionCallAction::Continue => CompletionCallOutcome::Proceed(None),
-    }
-}
-
-/// Append a finished run's messages to conversation memory, logging and
-/// proceeding on failure. Shared `Done`-arm behavior for both drivers.
-pub(crate) async fn append_run_messages(
-    memory_handle: Option<&(Arc<dyn ConversationMemory>, rig_core::id::ConversationId)>,
-    messages: &[Message],
-) {
-    // Clone into an owned vec only when there is a backend to append to — the
-    // common no-memory path pays nothing.
-    if let Some((memory, id)) = memory_handle
-        && let Err(err) = memory.append(id, messages.to_vec()).await
-    {
-        tracing::warn!(
-            error = %err,
-            conversation_id = %id,
-            "conversation memory append failed; surfacing final response anyway"
-        );
-    }
-}
-
-/// Whether (and how) a tool call executed, for [`run_single_tool`].
-pub(crate) enum ToolExecution {
-    /// The tool's body ran. Carries the **effective** tool call — the model's
-    /// call with any [`ToolCallAction::Rewrite`] hook
-    /// rewrite applied — so the driver can surface it in the
-    /// [`ToolExecutionCommitted`](crate::agent::prompt_request::streaming::MultiTurnStreamItem::ToolExecutionCommitted)
-    /// event (what actually ran, not the model's original arguments). Boxed to
-    /// keep this enum small (a `ToolCall` is large next to the empty `Skipped`).
-    Executed(Box<ToolCall>),
-    /// A tool-call hook returned [`ToolCallAction::Skip`]: the
-    /// body did not run, so no execution-commit is surfaced — but the skip result
-    /// is still delivered to the model (and surfaced as a `ToolResult`).
-    Skipped,
-}
-
-/// Outcome of [`run_single_tool`]: the tool-result content plus whether the
-/// tool's body ran (and the effective call) or a hook skipped it.
-pub(crate) struct ToolCallOutcome {
-    /// The tool result delivered to the model (a real output, a redacted
-    /// replacement, or a hook skip reason).
-    pub content: UserContent,
-    /// How the call resolved: executed (with the effective tool call) or skipped.
-    pub execution: ToolExecution,
-}
-
-/// Execute a single tool call, firing the `ToolCall` and `ToolResult` hooks and
-/// shaping the result. **Shared by the blocking and streaming drivers** so a
-/// tool call behaves identically in both: same hook events, same fail-closed
-/// skip/terminate handling, and the same result shaping. Hook skips become
-/// [`ToolResult::skipped`], and every result is converted directly into typed
-/// message content through [`tool_result_output`] without reparsing text.
-/// Records `gen_ai.tool.*` on the current span;
-/// `error_history` builds a cancellation error if a hook terminates the run.
-/// Returns whether the tool body executed via [`ToolCallOutcome::execution`].
-pub(crate) async fn run_single_tool(
-    runner: &AgentRunner,
-    ctx: &HookContext,
-    tool_snapshot: &ToolRegistrySnapshot,
-    tool_call: &ToolCall,
-    internal_call_id: rig_core::id::InternalCallId,
-    error_history: &[Message],
-) -> Result<ToolCallOutcome, PromptError> {
-    let hooks = &runner.config.hooks;
-    let tool_context = &runner.tool_context;
-    let record_content = runner.config.record_telemetry_content;
-    let tool_name = &tool_call.function.name;
-    // `mut` so a tool-call hook can rewrite the arguments the tool
-    // runs with (the model's emitted arguments are otherwise used verbatim).
-    let mut args = json_utils::serialize_json_value(&tool_call.function.arguments);
-
-    let tool_span = tracing::Span::current();
-    tool_span.record("gen_ai.tool.name", tool_name);
-    tool_span.record("gen_ai.tool.call.id", tool_call.id.as_str());
-    if record_content {
-        tool_span.record("gen_ai.tool.call.arguments", &args);
-    }
-
-    // Resolve the `ToolCall` hook chain. A proceeding chain carries any
-    // `ToolCallAction::Rewrite` in the action itself; a chain that a later hook
-    // short-circuits with `Skip`/`Stop` salvages the accumulated
-    // rewrite into `salvaged_rewrite` so it is *not* lost — the rewritten args
-    // must still be reported on the skipped `ToolResult` and in tracing rather
-    // than leaking the model's original args (see [`HookStack::resolve_tool_call`]).
-    let (action, salvaged_rewrite) = hooks
-        .resolve_tool_call(
-            ctx,
-            ToolCallEvent {
-                tool_name,
-                tool_call_id: Some(tool_call.id.as_str()),
-                internal_call_id,
-                args: &args,
-            },
-        )
-        .await;
-
-    // Apply a salvaged rewrite (short-circuit path only) so `args` — what the
-    // `ToolResult` reports — and the span reflect the effective arguments.
-    if let Some(rewritten) = salvaged_rewrite.as_ref() {
-        args = json_utils::serialize_json_value(rewritten);
-        if record_content {
-            tool_span.record("gen_ai.tool.call.arguments", &args);
-        }
-        tracing::debug!(
-            tool_name = tool_name,
-            "tool-call arguments rewritten by a hook"
-        );
-    }
-
-    // On `Skip` the body does not run and the structured outcome is `Skipped`;
-    // otherwise the tool executes into a structured `ToolResult`.
-    // `effective_args` is what the tool actually ran with (the model's, a hook's
-    // `ToolCallAction::Rewrite` replacement, or a salvaged rewrite) — surfaced in the
-    // execution-commit event so a redaction rewrite does not leak. Unused for a skip.
-    let mut skipped: Option<ToolResult> = None;
-    let effective_args: serde_json::Value = match action {
-        ToolCallAction::Stop(reason) => {
-            return Err(PromptError::prompt_cancelled(
-                error_history.to_vec(),
-                reason,
-            ));
-        }
-        ToolCallAction::Skip(reason) => {
-            tracing::info!(tool_name = tool_name, reason = reason, "Tool call rejected");
-            // Synthetic rejection: `Skipped` outcome, message delivered verbatim.
-            // Still fires the `ToolResult` hook so a policy observes the skip.
-            skipped = Some(ToolResult::skipped(reason));
-            // A skip runs nothing; its effective args are the salvaged rewrite
-            // (if any) so tracing/history stay consistent, though they go unused.
-            salvaged_rewrite.unwrap_or_else(|| tool_call.function.arguments.clone())
-        }
-        ToolCallAction::Rewrite(replacement) => {
-            // Proceeding rewrite: re-record the span so the trace, and the
-            // downstream `ToolResult` event, reflect what the tool actually
-            // received rather than what the model emitted.
-            args = json_utils::serialize_json_value(&replacement);
-            if record_content {
-                tool_span.record("gen_ai.tool.call.arguments", &args);
-            }
-            tracing::debug!(
-                tool_name = tool_name,
-                "tool-call arguments rewritten by a hook"
-            );
-            replacement
-        }
-        ToolCallAction::Run => tool_call.function.arguments.clone(),
-    };
-
-    // Resolve the structured execution result and how the call surfaced. A skip
-    // produces no execution-commit event; a real execution carries the effective
-    // tool call (the model's call with any `ToolCallAction::Rewrite` applied).
-    let (exec, execution, dispatch_context) = match skipped {
-        Some(exec) => (exec, ToolExecution::Skipped, tool_context.for_dispatch()),
-        None => {
-            let mut effective_tool_call = tool_call.clone();
-            effective_tool_call.function.arguments = effective_args;
-            let ToolDispatch {
-                result: exec,
-                context: dispatch_context,
-            } = tool_snapshot.dispatch(tool_name, &args, tool_context).await;
-            (
-                exec,
-                ToolExecution::Executed(Box::new(effective_tool_call)),
-                dispatch_context,
-            )
-        }
-    };
-    // Presentation rewrites happen after execution. The raw structured result
-    // and per-dispatch context remain unchanged for every hook.
-    let result_action = hooks
-        .on_tool_result(
-            ctx,
-            ToolResultEvent {
-                tool_name,
-                tool_call_id: Some(tool_call.id.as_str()),
-                internal_call_id,
-                args: &args,
-                presentation: exec.output(),
-                raw_result: &exec,
-                tool_context: &dispatch_context,
-            },
-        )
-        .await;
-    // Outcome metadata describes the execution itself, while result content
-    // follows the same presentation policy as the model. This keeps redaction
-    // and stop hooks from leaking raw tool output through telemetry.
-    record_tool_result(&tool_span, &exec);
-
-    let result_content = match result_action {
-        ToolResultAction::Stop(reason) => {
-            return Err(PromptError::prompt_cancelled(
-                error_history.to_vec(),
-                reason,
-            ));
-        }
-        ToolResultAction::Rewrite(replacement) => {
-            if record_content {
-                tool_span.record("gen_ai.tool.call.result", replacement.render());
-            }
-            replacement
-        }
-        ToolResultAction::Keep => {
-            if record_content {
-                tool_span.record("gen_ai.tool.call.result", exec.output().render());
-            }
-            exec.output().clone()
-        }
-    };
-    let content = tool_result_output(
-        tool_call.id.clone(),
-        tool_call.provider.clone(),
-        tool_call.function.name.clone(),
-        result_content,
-    );
-    Ok(ToolCallOutcome { content, execution })
-}
-
-fn record_tool_result(span: &tracing::Span, result: &ToolResult) {
-    span.record("gen_ai.tool.call.outcome", result.status_name());
-    if let Some(error) = result.error() {
-        span.record("gen_ai.tool.error.type", error.kind().as_str());
-    }
-}
-
-/// Build the per-tool `execute_tool` span carrying the `gen_ai.tool.*` fields
-/// that [`run_single_tool`] records on the current span. Parented to the
-/// contextual current span; the blocking driver additionally chains it via
-/// `follows_from`, while the streaming driver uses it as-is. Shared by both
-/// drivers so the span shape stays defined in one place.
-pub(crate) fn new_execute_tool_span() -> tracing::Span {
-    info_span!(
-        "execute_tool",
-        gen_ai.operation.name = "execute_tool",
-        gen_ai.tool.type = "function",
-        gen_ai.tool.name = tracing::field::Empty,
-        gen_ai.tool.call.id = tracing::field::Empty,
-        gen_ai.tool.call.arguments = tracing::field::Empty,
-        gen_ai.tool.call.result = tracing::field::Empty,
-        gen_ai.tool.call.outcome = tracing::field::Empty,
-        gen_ai.tool.error.type = tracing::field::Empty
-    )
-}
-
-/// [`TurnSource`] for the blocking surface: each turn issues a unary
-/// `model.completion()` request and feeds the whole response into the machine.
-/// Emits no intermediate items (the blocking surface folds the engine to its
-/// final response), but keeps the blocking driver's linear `follows_from` span
-/// chain across chat and tool spans.
-pub(crate) struct UnaryTurnSource {
-    /// Sequences chat and tool spans into a linear `follows_from` chain (the
-    /// streaming surface parents into a tree instead and does not chain).
-    ///
-    /// Atomic rather than `Cell` despite being driven by a single sequential
-    /// task: `run_tool_calls` passes `chain_span` as a closure into
-    /// `drive_tool_calls`, whose returned `DriveStream` is `Send`. That makes the
-    /// closure capture `&self`, so `&UnaryTurnSource` must be `Send`, i.e.
-    /// `UnaryTurnSource: Sync` — which `AtomicU64` provides and `Cell` does not.
-    current_span_id: AtomicU64,
-    record_telemetry_content: bool,
-}
-
-impl UnaryTurnSource {
-    pub(crate) fn new(record_telemetry_content: bool) -> Self {
-        Self {
-            current_span_id: AtomicU64::new(0),
-            record_telemetry_content,
-        }
-    }
-
-    /// Chain `span` onto the previous step's span and record it as the new chain
-    /// head, preserving the blocking driver's linear causal trace.
-    fn chain_span(&self, span: tracing::Span) -> tracing::Span {
-        let span = match self.current_span_id.load(Ordering::Relaxed) {
-            0 => span,
-            id => {
-                span.follows_from(Id::from_u64(id));
-                span
-            }
-        };
-        if let Some(id) = span.id() {
-            self.current_span_id.store(id.into_u64(), Ordering::Relaxed);
-        }
-        span
-    }
-}
-
-impl TurnSource for UnaryTurnSource {
-    fn open_chat_span(
-        &self,
-        runner: &AgentRunner,
-        effective_preamble: Option<&str>,
-    ) -> tracing::Span {
-        let chat_span = build_chat_span!(runner, effective_preamble, "chat", "chat");
-        self.chain_span(chat_span)
-    }
-
-    fn run_model_turn<'a>(
-        &'a mut self,
-        runner: &'a AgentRunner,
-        hook_ctx: &'a HookContext,
-        run: &'a mut AgentRun,
-        prepared: PreparedCompletionRequest,
-        chat_span: tracing::Span,
-        _agent_span: &'a tracing::Span,
-        current_prompt: Message,
-    ) -> DriveStream<'a> {
-        Box::pin(async_stream::stream! {
-            // Content telemetry for the accepted provider turn. Called at each
-            // terminal site (stop, terminate, accept) rather than hoisted: a
-            // retried turn must not record output for the discarded attempt.
-            let record_accepted_turn = |run: &AgentRun| {
-                if runner.config.record_telemetry_content
-                    && let Some(choice) = run.accepted_turn_choice()
-                {
-                    rig_core::telemetry::record_model_output(&chat_span, &choice, true);
-                }
-            };
-
-            // Bound before the builder is consumed: this is the cap this exact
-            // attempt was prepared with, patches included, and it is what the
-            // per-turn hook reports. Reading it later off the agent config would
-            // silently drop a completion-call hook's patch.
-            let attempt_max_tokens = prepared.max_tokens;
-
-            let resp = match prepared.builder.send().instrument(chat_span.clone()).await {
-                Ok(resp) => resp,
-                Err(err) => {
-                    yield Err(StreamingError::from(err));
-                    return;
-                }
-            };
-
-            // Normalized once, then shared by run state and the per-turn hook, so
-            // the two cannot report different reasons for one attempt.
-            let attempt_finish_reason = resp.finish_reason();
-
-            let mut outcome = match run.model_response(ModelTurn::from_response_parts(
-                &resp,
-                prepared.executable_tool_names,
-                prepared.allowed_tool_names,
-            )) {
-                Ok(outcome) => outcome,
-                Err(err) => {
-                    yield Err(Box::new(err).into());
-                    return;
-                }
-            };
-
-            loop {
-                match outcome {
-                    ModelTurnOutcome::NeedsResolution(context) => {
-                        let action = runner
-                            .config.hooks
-                            .on_invalid_tool_call(hook_ctx, &context)
-                            .await;
-                        let resolution = match action {
-                            Some(action) => run.resolve_invalid_tool_call(action),
-                            None
-                                if runner.unhandled_invalid_tool_call_policy
-                                    == UnhandledInvalidToolCallPolicy::IgnoreForExtractor =>
-                            {
-                                run.ignore_invalid_tool_call()
-                            }
-                            None => run.resolve_invalid_tool_call(InvalidToolCallAction::fail()),
-                        };
-                        outcome = match resolution {
-                            Ok(outcome) => outcome,
-                            Err(err) => {
-                                yield Err(Box::new(err).into());
-                                return;
-                            }
-                        };
-                    }
-                    ModelTurnOutcome::TurnRetried => break,
-                    ModelTurnOutcome::Continue {
-                        response_hook_suppressed,
-                    } => {
-                        if !response_hook_suppressed {
-                            let identity = resp.identity();
-                            let settlement = settle_model_turn(
-                                &runner.config.hooks,
-                                hook_ctx,
-                                run,
-                                AssembledTurn {
-                                    prompt: &current_prompt,
-                                    content: &resp.choice,
-                                    usage: resp.usage,
-                                    identity: &identity,
-                                    finish_reason: attempt_finish_reason.as_ref(),
-                                    max_tokens: attempt_max_tokens,
-                                    raw: &resp.raw,
-                                },
-                            )
-                            .await;
-                            match settlement {
-                                Ok(ModelTurnDecision::Advance) => {}
-                                Ok(ModelTurnDecision::Retried) => break,
-                                Ok(ModelTurnDecision::Terminate(reason)) => {
-                                    record_accepted_turn(run);
-                                    yield Err(StreamingError::Prompt(Box::new(
-                                        run.cancel_error(reason),
-                                    )));
-                                    return;
-                                }
-                                Err(err) => {
-                                    yield Err(StreamingError::Prompt(Box::new(err)));
-                                    return;
-                                }
-                            }
-                        }
-
-                        record_accepted_turn(run);
-                        break;
-                    }
-                }
-            }
-        })
-    }
-
-    fn run_tool_calls<'a>(
-        &'a self,
-        runner: &'a AgentRunner,
-        hook_ctx: &'a HookContext,
-        run: &'a mut AgentRun,
-        calls: Vec<PendingToolCall>,
-        tool_snapshot: Arc<ToolRegistrySnapshot>,
-    ) -> DriveStream<'a> {
-        // The blocking surface chains tool spans into its linear `follows_from`
-        // sequence (chat -> tool -> chat), and discards the yielded items, so it
-        // skips building them.
-        drive_tool_calls(
-            runner,
-            hook_ctx,
-            run,
-            calls,
-            tool_snapshot,
-            |span| self.chain_span(span),
-            false,
-        )
-    }
-
-    fn record_run_level_telemetry(
-        &self,
-        agent_span: &tracing::Span,
-        response: &PromptResponse,
-        created_agent_span: bool,
-    ) {
-        // Record run-level completion + usage onto the agent span, but only when
-        // we created it — never pollute a caller-supplied outer span. The usage
-        // fields go through the same recorder the streaming surface uses; the
-        // blocking surface additionally records the final completion text.
-        if created_agent_span {
-            if self.record_telemetry_content {
-                agent_span.record("gen_ai.completion", &response.output);
-            }
-            agent_span.record_token_usage(&response.usage);
-        }
-    }
-
-    fn final_item(&self, _response: &PromptResponse) -> Option<MultiTurnStreamItem> {
-        // The blocking surface folds the engine and discards the final item, so
-        // building it (an extra full-response clone) is skipped entirely.
-        None
-    }
-}
-
 impl AgentRunner {
     pub(crate) async fn run_with_error_usage(
         mut self,
@@ -1169,7 +496,3 @@ impl AgentRunner {
 
 #[cfg(test)]
 mod tests;
-
-#[cfg(test)]
-#[allow(irrefutable_let_patterns, unreachable_patterns)]
-mod migrated_tests;
